% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_zed_indices.R
\name{area_zed_indices}
\alias{area_zed_indices}
\title{Area Zed-Indices}
\usage{
area_zed_indices()
}
\arguments{
\item{API_key}{An API key from Zoopla.}
}
\value{
A list.
}
\description{
Area Zed-Indices
}
\examples{

}
\references{
\url{http://developer.zoopla.com/docs/read/area_zed_indices}
}
