% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_appraisals.R
\name{arrange_appraisals}
\alias{arrange_appraisals}
\title{Arrange Appraisals}
\usage{
arrange_appraisals()
}
\arguments{
\item{API_key}{An API key from Zoopla.}
}
\value{
A list.
}
\description{
Arrange Appraisals
}
\examples{

}
\references{
\url{http://developer.zoopla.com/docs/read/Arrange_Appraisals}
}
