% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_area_sold_price.R
\name{average_area_sold_price}
\alias{average_area_sold_price}
\title{Average Area Sold Price}
\usage{
average_area_sold_price(area = NULL, output = NULL, API_key = NULL)
}
\arguments{
\item{area}{A string representing desired area. Could be a postcode, outcode, or a name. Should unambiguous, otherwise the API will return an error.}

\item{output}{A string representing desired output: "outcode", "town", "county", "country". This should cover a larger area than the 'area' variable.}

\item{API_key}{An API key from Zoopla.}
}
\value{
A list.
}
\description{
Average Area Sold Price
}
\examples{
\dontrun{

average_area_sold_price("Edinburgh", "country", "YOUR_API_KEY")
average_area_sold_price("EH1 2NG", "outcode", "YOUR_API_KEY")
average_area_sold_price("EH1", "county", "YOUR_API_KEY")
}

}
\references{
\url{http://developer.zoopla.com/docs/read/Average_area_sold_price}
}
