% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_sold_prices.R
\name{average_sold_prices}
\alias{average_sold_prices}
\title{Average Sold Prices}
\usage{
average_sold_prices(postcode = NULL, output = NULL, area_type = NULL,
  API_key = NULL)
}
\arguments{
\item{postcode}{A string representing a valid UK outcode or postcode, e.g. "NW10" or "EH1 2NG".}

\item{output}{A string representing desired output: "outcode" or "county". "county" can be combined with "streets", "towns" or "outcodes" in the area_type. "outcode" can be combined with "postcodes" or "streets" in the area_type.}

\item{area_type}{A string representing desired area: "streets", "towns", "outcodes" or "postcodes".}

\item{API_key}{An API key from Zoopla.}
}
\value{
A list.
}
\description{
Average Sold Prices
}
\examples{
\dontrun{

average_sold_prices("EH1 2NG", "outcode", "streets", "YOUR_API_KEY")
average_sold_prices("CF10 3XQ", "county", "outcodes", "YOUR_API_KEY")
}

}
\references{
\url{http://developer.zoopla.com/docs/read/Average_Sold_Prices}
}
