% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_autocomplete.R
\name{geo_autocomplete}
\alias{geo_autocomplete}
\title{Geo Autocomplete}
\usage{
geo_autocomplete(search_term = NULL, search_type = "listings",
  API_key = NULL)
}
\arguments{
\item{search_term}{A string holding the search term to be auto completed.}

\item{search_type}{A string. Allowed values: "listings" or "property". Defaults to "listings".}

\item{API_key}{An API key from Zoopla.}
}
\value{
A list
}
\description{
Geo Autocomplete
}
\examples{
geo_autocomplete("ruislip", "listings", "YOUR_API_KEY")

}
\references{
\url{http://developer.zoopla.com/docs/read/Geo_Autocomplete}
}
