% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_session_id.R
\name{get_session_id}
\alias{get_session_id}
\title{Get Session ID}
\usage{
get_session_id(API_key = NULL)
}
\arguments{
\item{API_key}{An API key from Zoopla.}
}
\value{
A string - session ID.
}
\description{
Get Session ID
}
\examples{
get_session_id("YOUR_API_KEY")

}
\references{
\url{http://developer.zoopla.com/docs/read/Get_Session_ID}
}
