% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_info_graphs.R
\name{local_info_graphs}
\alias{local_info_graphs}
\title{Local Info Graphs}
\usage{
local_info_graphs()
}
\arguments{
\item{API_key}{An API key from Zoopla.}
}
\value{
A list.
}
\description{
Local Info Graphs
}
\examples{

}
\references{
\url{http://developer.zoopla.com/docs/read/Local_Info_Graphs}
}
