% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property_listings.R
\name{property_listings}
\alias{property_listings}
\title{Property Listings}
\usage{
property_listings()
}
\arguments{
\item{API_key}{An API key from Zoopla.}
}
\value{
A list.
}
\description{
Property Listings
}
\examples{

}
\references{
\url{http://developer.zoopla.com/docs/read/Property_listings}
}
