% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property_rich_list.R
\name{property_rich_list}
\alias{property_rich_list}
\title{Property Rich List}
\usage{
property_rich_list()
}
\arguments{
\item{API_key}{An API key from Zoopla.}
}
\value{
A list.
}
\description{
Property Rich List
}
\examples{

}
\references{
\url{http://developer.zoopla.com/docs/read/Property_Rich_List}
}
