% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_estimate.R
\name{refine_estimate}
\alias{refine_estimate}
\title{Refine Estimate}
\usage{
refine_estimate(API_key, session, property_id, property_type, tenure,
  num_bedrooms, num_bathrooms, num_receptions)
}
\arguments{
\item{API_key}{An API key from Zoopla.}
}
\value{
A list.
}
\description{
Refine Estimate
}
\examples{
\dontrun{

sid <- get_session_id(APIKEY)

# if the values provided are incorrect then error 400 is returned
refine_estimate()
}

}
\references{
\url{http://developer.zoopla.com/docs/read/Refine_Estimate}
}
