% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_zed_indices.R
\name{area_zed_indices}
\alias{area_zed_indices}
\title{Area Zed-Indices}
\usage{
area_zed_indices(ordering = "descending", postcode = NULL,
  area_type = NULL, output_type = NULL, page_number = NULL,
  page_size = NULL, API_key = NULL)
}
\arguments{
\item{API_key}{An API key from Zoopla.}
}
\value{
A list.
}
\description{
Area Zed-Indices
}
\examples{
\dontrun{

area_zed_indices("descending", "sw185rw", "postcodes", "outcode", 1, 10, "YOUR_API_KEY")
area_zed_indices(postcode = "sw185rw", area_type = "postcodes", output_type = "outcode", API_key = "YOUR_API_KEY")
}

}
\references{
\url{http://developer.zoopla.com/docs/read/area_zed_indices}
}
