% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_viewing.R
\name{arrange_viewing}
\alias{arrange_viewing}
\title{Arrange Viewing}
\usage{
arrange_viewing(API_key = NULL)
}
\arguments{
\item{API_key}{An API key from Zoopla.}
}
\value{
A list.
}
\description{
Arrange Viewing
}
\examples{
\dontrun{

arrange_viewing()
}

}
\references{
\url{http://developer.zoopla.com/docs/read/Arrange_Viewing}
}
