% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_info_graphs.R
\name{local_info_graphs}
\alias{local_info_graphs}
\title{Local Info Graphs}
\usage{
local_info_graphs(area = NULL, API_key = NULL)
}
\arguments{
\item{area}{A string representing a valid UK outcode, e.g. "W12".}

\item{API_key}{An API key from Zoopla.}
}
\value{
A list.
}
\description{
Generate a set of graphs of local info for an outcode (and optional incode) and return the URL to the generated image.
}
\examples{
\dontrun{

local_info_graphs("W12", "YOUR_API_KEY")
}

}
\references{
\url{http://developer.zoopla.com/docs/read/Local_Info_Graphs}
}
