% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode_properties.R
\name{postcode_properties}
\alias{postcode_properties}
\title{Postcode Properties}
\usage{
postcode_properties(postcode = NULL, API_key = NULL)
}
\arguments{
\item{postcode}{A string representing a valid UK postcode, e.g. "EH1 2NG".}

\item{API_key}{An API key from Zoopla.}
}
\value{
A list.
}
\description{
Postcode Properties
}
\examples{
\dontrun{

postcode_properties("EH12NG", "YOUR_API_KEY")
postcode_properties("EH1 2NG", "YOUR_API_KEY")
}

}
\references{
\url{https://developer.zoopla.co.uk/docs/read/Arrange_Appraisals}
}
