% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property_listings.R
\name{property_listings}
\alias{property_listings}
\title{Property Listings}
\usage{
property_listings(postcode = NULL, area = NULL, API_key = NULL)
}
\arguments{
\item{postcode}{A string. Valid UK postcode or outcode.}

\item{area}{A string. Area of interest, e.g. "Nottingham".}

\item{API_key}{An API key from Zoopla.}
}
\value{
A list.
}
\description{
Property Listings
}
\examples{
\dontrun{

property_listings(postcode = "E1", API_key = APIKEY)
property_listings(area = "Nottingham", API_key = APIKEY)
}

}
\references{
\url{http://developer.zoopla.com/docs/read/Property_listings}
}
