% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_estimate.R
\name{refine_estimate}
\alias{refine_estimate}
\title{Refine Estimate}
\usage{
refine_estimate(property_id, session, property_type, tenure, num_bedrooms,
  num_bathrooms, num_receptions, API_key)
}
\arguments{
\item{property_id}{Property id. Can be extracted with postcode_properties. Need to be a live listing.}

\item{session}{Session ID. Use get_session_id to fetch it.}

\item{property_type}{Type of property. Must be one of the following: detached, link_detached, semi_detached, terraced, flat, end_terrace, maisonette, mews, town_house, cottage, bungalow, farm_barn, park_home}

\item{tenure}{Tenure. Allowed values: freehold, leasehold, share_of_freehold.}

\item{num_bedrooms}{The number of bedrooms located at the property.}

\item{num_bathrooms}{The number of bathrooms located at the property.}

\item{num_receptions}{The number of receptions located at the property.}

\item{API_key}{An API key from Zoopla.}
}
\value{
A list.
}
\description{
Request a more accurate Zoopla.co.uk Zed-Index based on extra data provided.
}
\examples{
\dontrun{

s_id <- get_session_id(APIKEY)

# get the property id but make sure that the ad is still live
prop_id <- postcode_properties("EH12NG", "YOUR_API_KEY")
p_id <- prop_id$agent$property_id

# if the values provided are incorrect then error 400 is returned
refine_estimate(p_id, s_id, "detached", "freehold", "4", "1", "1", APIKEY)
# refine_estimate("965889", "839ddfa0fa5d810d9afa148960775007",
"detached", "freehold", "4", "1", "1", APIKEY)
}

}
\references{
\url{http://developer.zoopla.com/docs/read/Refine_Estimate}
}
