% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_estimate.R
\name{refine_estimate}
\alias{refine_estimate}
\title{Refine Estimate}
\usage{
refine_estimate()
}
\arguments{
\item{API_key}{An API key from Zoopla.}
}
\value{
A list.
}
\description{
Refine Estimate
}
\examples{
\dontrun{

refine_estimate()
}

}
\references{
\url{http://developer.zoopla.com/docs/read/Refine_Estimate}
}
